// gitconnect-v2-next/types/weaviate.ts
export interface RepoLicense {
  name: string;
}

export interface WeaviateRepoUploadData {
  name: string;
  username?: string;
  description: string | null;
  tags?: string[];
  repoid: number;
  readme?: string | null;
  html_url?: string;
  license?: RepoLicense | string | null;
  forks_count?: number;
  stargazers_count?: number;
  watchers_count?: number;
  open_issues_count?: number;
  language?: string | null;
  mainLanguage?: string | null;
  languages_url?: string;
  languagePercententages?: string[] | null;
}